Add a guess at theoretical SNR
This adds a very naive guess at the theoretical SNR as a function
of magnitude.  It uses the sims_photUtils package to calculate
the theoretical counts.  The assumptions are:
1. flat spectrum source (zero AB color)
2. SNR at 24.5 in r is 5.
3. The systematic floor is 2%.

This is *not* a fit.  These values were picked as reasonable given
the data and what we expect from the system.